refactor: extract Keyword and Profile classes
refactor: extract the Keyword and the Profile classes from reactormodel.
