+++ src/ansys/chemkin/core/reactormodel.py
@@ -24,7 +24,6 @@
 
 import copy
 import ctypes
-from ctypes import c_double, c_int
 from pathlib import Path
 from typing import Any, Union
 
@@ -40,663 +39,18 @@ from ansys.chemkin.core.chemistry import (
     verbose,
 )
 from ansys.chemkin.core.color import Color
-from ansys.chemkin.core.constants import P_ATM
 from ansys.chemkin.core.inlet import Stream
+from ansys.chemkin.core.keyword import (
+    BooleanKeyword,
+    IntegerKeyword,
+    Keyword,
+    RealKeyword,
+    StringKeyword,
+)
 from ansys.chemkin.core.logger import logger
 from ansys.chemkin.core.mixture import Mixture
+from ansys.chemkin.core.profile import Profile
 from ansys.chemkin.core.surface_chemistry_controller import SurfaceChemistryController
-
-
-#
-# Base class for keyword data
-#
-class Keyword:
-    """A Chemkin style keyword."""
-
-    # supported Chemkin keyword data types
-    _keyworddatatypes = ["bool", "int", "float", "str"]
-    _valuetypes = (bool, int, float, str)
-    # required keywords that are given as reactor properties or as mixture properties
-    # and will be set by using the KINAll0D_SetupBatchInputs call
-    _protectedkeywords = [
-        "CONP",
-        "CONV",
-        "TRAN",
-        "STST",
-        "TGIV",
-        "ENRG",
-        "PRES",
-        "TEMP",
-        "TAU",
-        "TIME",
-        "XEND",
-        "FLRT",
-        "VDOT",
-        "SCCM",
-        "DIAM",
-        "REAC",
-        "GAS",
-        "INIT",
-        "XEST",
-        "SURF",
-        "ACT",
-        "TINL",
-        "FUEL",
-        "OXID",
-        "PROD",
-        "ASEN",
-        "ATLS",
-        "RTLS",
-        "EPST",
-        "EPSS",
-    ]
-    gasspecieskeywords = ["REAC", "XEST", "FUEL", "OXID"]
-    flowratekeywords = ["FLRT", "VDOT", "VEL", "SCCM"]
-    profilekeywords = [
-        "TPRO",
-        "PPRO",
-        "VPRO",
-        "QPRO",
-        "AINT",
-        "AEXT",
-        "DPRO",
-        "FPRO",
-        "SCCMPRO",
-        "VDOTPRO",
-        "VELPRO",
-        "TINPRO",
-        "AFLO",
-    ]
-    fourspaces = "    "
-    # Under the default API-call mode, important keywords (the _protectedkeywords)
-    # are set by direct API calls, the rest of the keywords can be set by
-    # keyword input lines (i.e., using the setkeyword method).
-    # Under the full-keyword mode, all keywords and their parameters are set
-    # by keyword input lines, and specifying those _protectedkeywords via
-    # the setkeyword method are required.
-    no_fullkeyword = True  # default: API-call mode
-
-    def __init__(self, phrase: str, value: Union[float, bool, str], data_type: str):
-        """Initialize the Chemkin keyword."""
-        """
-        Initialize the Chemkin keyword.
-
-        Parameters
-        ----------
-            phrase: string
-                Chemkin keyword phrase
-            value: indicated by the data_type, {int, float, string, bool}
-                value assigned to the Chemkin keyword
-            data_type: string, {'int', 'float', 'string', or 'bool'}
-                data type of value
-
-        """
-        self._set = False
-        ierr = 0
-        # check value data type
-        if data_type not in Keyword._keyworddatatypes:
-            # the declared data type is not supported
-            msg = [
-                Color.PURPLE,
-                "unsupported data type specified",
-                data_type,
-                Color.END,
-            ]
-            this_msg = Color.SPACE.join(msg)
-            logger.error(this_msg)
-            if not isinstance(value, (bool, int, float, str)):
-                # value does not match the declared data type
-                msg = [
-                    Color.PURPLE,
-                    "variable has different data type",
-                    str(type(value)),
-                    Color.END,
-                ]
-                this_msg = Color.SPACE.join(msg)
-                logger.error(this_msg)
-            ierr = 1
-        # block the protected keywords
-        if Keyword.no_fullkeyword:
-            if phrase.upper() in Keyword._protectedkeywords:
-                msg = [
-                    Color.PURPLE,
-                    "use reactor property setter to assign",
-                    phrase,
-                    "value\n",
-                    Color.SPACEx6,
-                    "for example, to set the reactor volume use:",
-                    '"MyBatchReactor.volume = 100"',
-                    Color.END,
-                ]
-                this_msg = Color.SPACE.join(msg)
-                logger.error(this_msg)
-                ierr = 2
-        if ierr > 0:
-            return
-        self._key = phrase  # Chemkin keyword phrase
-        self._value = value  # value assigned to the keyword
-        self._data_type = data_type  # data type of the values
-        self._prefix = ""  # a prefix to the keyword that can be used
-        # to comment out/disable the keyword by setting it to '!'
-        self._set = True
-
-    @staticmethod
-    def setfullkeywords(mode: bool):
-        """Require all keywords and their parameters must be specified."""
-        """All keywords and their parameters must be specified by
-        using the setkeyword method and will be passed to the reactor model
-        for further processing.
-
-        Parameters
-        ----------
-            mode: boolean
-                turn the full keyword mode ON/OFF
-
-        """
-        if mode:
-            # turn ON the full keyword mode (no checking on protected keywords)
-            Keyword.no_fullkeyword = False
-        else:
-            Keyword.no_fullkeyword = True
-
-    def show(self):
-        """Display the Chemkin keyword and its parameter value."""
-        if self._set:
-            if isinstance(self._value, (int, float)):
-                msg = [
-                    Color.YELLOW,
-                    "keyword",
-                    "'" + self._key + "':",
-                    "value =",
-                    str(self._value),
-                    Color.END,
-                ]
-                this_msg = Color.SPACE.join(msg)
-                logger.info(this_msg)
-            elif isinstance(self._value, bool):
-                if self._value:
-                    msg = [
-                        Color.YELLOW,
-                        "keyword",
-                        "'" + self._key + "':",
-                        "value = True",
-                        Color.END,
-                    ]
-                    this_msg = Color.SPACE.join(msg)
-                    logger.info(this_msg)
-                else:
-                    msg = [
-                        Color.YELLOW,
-                        "keyword",
-                        "'" + self._prefix + self._key + "':",
-                        "value = Disabled",
-                        Color.END,
-                    ]
-                    this_msg = Color.SPACE.join(msg)
-                    logger.info(this_msg)
-            else:
-                msg = [
-                    Color.YELLOW,
-                    "keyword",
-                    "'" + self._key + "':",
-                    "value =",
-                    str(self._value),
-                    Color.END,
-                ]
-                this_msg = Color.SPACE.join(msg)
-                logger.info(this_msg)
-        else:
-            msg = [
-                Color.YELLOW,
-                "keyword",
-                "'" + self._key + "':",
-                "value not set.",
-                Color.END,
-            ]
-            this_msg = Color.SPACE.join(msg)
-            logger.info(this_msg)
-
-    def resetvalue(self, value: Union[float, bool, str]):
-        """Reset the parameter value of an existing keyword."""
-        """
-        Reset the parameter value of an existing keyword.
-
-        Parameters
-        ----------
-            value: indicated by the data_type, {int, float, string, bool}
-                keyword parameter
-
-        """
-        if isinstance(value, Keyword._valuetypes):
-            if isinstance(value, bool):
-                if value:
-                    # true: keep the keyword active
-                    self._prefix = ""
-                    self._value = True
-                else:
-                    # false: disable the keyword
-                    self._prefix = "!"
-                    self._value = False
-            else:
-                # integer, float, or string parameter
-                self._value = value
-        else:
-            msg = [
-                Color.PURPLE,
-                "value has a wrong data type",
-                type(value),
-                "value will not be reset.\n",
-                Color.SPACEx6,
-                "data type expected by keyword",
-                self._key,
-                "is",
-                self._data_type,
-                Color.END,
-            ]
-            this_msg = Color.SPACE.join(msg)
-            logger.error(this_msg)
-
-    def parametertype(self) -> type:
-        """Get parameter type of the keyword."""
-        """
-        Get parameter type of the keyword.
-
-        Returns
-        -------
-            parameter_data_type: string, {'int', 'float', 'string', 'bool'}
-                parameter data type
-
-        """
-        return type(self._value)
-
-    @property
-    def value(self) -> Union[int, float, bool, str]:
-        """Get parameter value of the keyword."""
-        """
-        Get parameter value of the keyword.
-
-        Returns
-        -------
-            parameter value: {integer, floating, string, or boolean}
-
-        """
-        # extract the keyword value
-        if self._data_type == "bool":
-            mode = self._prefix != "!"
-            return mode
-        else:
-            return self._value
-
-    @property
-    def keyphrase(self) -> str:
-        """Get the phrase of the keyword."""
-        """
-        Get the phrase of the keyword.
-
-        Returns
-        -------
-            keyword phrase: string
-
-        """
-        return self._key
-
-    @property
-    def keyprefix(self) -> bool:
-        """Get the status of the keyword."""
-        """
-        Get the status of the keyword.
-
-        Returns
-        -------
-            status: boolean
-                keyword is ON/OFF
-
-        """
-        if self._prefix != "!":
-            return True
-        else:
-            return False
-
-    def getvalue_as_string(self) -> tuple[int, str]:
-        """Create the keyword input line for Chemkin applications."""
-        """
-        Create the keyword input line for Chemkin applications.
-
-        Returns
-        -------
-            linelength: integer
-                line length
-            line: string
-                keyword value
-
-        """
-        # initialization
-        line = ""
-        linelength = 0
-        # assembly the keyword line
-        if self._data_type == "bool":
-            # boolean keyword (active or disabled by '!')
-            line = self._prefix + self._key
-        else:
-            # integer, double, or string parameter
-            line = self._prefix + self._key + Keyword.fourspaces + str(self._value)
-
-        linelength = len(line)
-        return linelength, line
-
-
-#
-# This keyword type is used to distinguish keywords that act as
-# on/off switches by their presence
-#
-class BooleanKeyword(Keyword):
-    """Chemkin boolean keyword."""
-
-    def __init__(self, phrase: str):
-        """Set up a Chemkin keyword with a boolean parameter."""
-        """
-        Set up a Chemkin keyword with a boolean parameter or
-        with no parameter.
-
-        Parameters
-        ----------
-            phrase: string
-                Chemkin keyword phrase
-
-        """
-        value = True
-        super().__init__(phrase, value, "bool")
-
-
-#
-# This keyword type is used to hold integer keyword types
-# (not sure if there actually are any of these)
-#
-class IntegerKeyword(Keyword):
-    """A Chemkin integer keyword."""
-
-    def __init__(self, phrase: str, value: int = 0):
-        """Set up a Chemkin keyword with an integer parameter."""
-        """
-        Set up a Chemkin keyword with an integer parameter.
-
-        Parameters
-        ----------
-            phrase: string
-                Chemkin keyword phrase
-            value: integer
-                parameter value
-
-        """
-        super().__init__(phrase, value, "int")
-
-
-#
-# This keyword type is used to hold real keyword types
-#
-class RealKeyword(Keyword):
-    """A Chemkin real keyword."""
-
-    def __init__(self, phrase: str, value: float = 0.0e0):
-        """Set up a Chemkin keyword with a real number parameter."""
-        """Set up a Chemkin keyword with a real number
-        (floating number) parameter.
-
-        Parameters
-        ----------
-            phrase: string
-                Chemkin keyword phrase
-            value: double
-                parameter value
-
-        """
-        super().__init__(phrase, value, "float")
-
-
-#
-# This keyword type is used to hold string keyword types
-#
-class StringKeyword(Keyword):
-    """A Chemkin string keyword."""
-
-    def __init__(self, phrase: str, value: str = ""):
-        """Set up a Chemkin keyword with a string parameter."""
-        """Set up a Chemkin keyword with a string parameter.
-
-        Parameters
-        ----------
-            phrase: string
-                Chemkin keyword phrase
-            value: string
-                parameter value
-
-        """
-        if len(value) <= 0:
-            msg = [Color.PURPLE, "no string parameter given", Color.END]
-            this_msg = Color.SPACE.join(msg)
-            logger.error(this_msg)
-            exit()
-        super().__init__(phrase, value, "str")
-
-
-class Profile:
-    """Chemkin profile keyword class."""
-
-    def __init__(
-        self,
-        key: str,
-        x: npt.NDArray[np.double],
-        y: npt.NDArray[np.double],
-        label: bool = False,
-    ):
-        """Create a profile object."""
-        """Create a profile object.
-
-        Parameters
-        ----------
-            key: string
-                profile keyword
-            x: 1-D double array
-                position of the profile data points
-            y: 1-D double array
-                variable value of the profile data
-            label: boolean, optional
-                True: key includes a reactor/species/inlet name
-
-        """
-        # initialization
-        self._profilekeyword = ""
-        self._status = 0
-        # check
-        if key in Keyword.profilekeywords:
-            self._profilekeyword = key
-        elif label:
-            this_key = key.split()[0]
-            if this_key in Keyword.profilekeywords:
-                self._profilekeyword = key
-        else:
-            msg = [
-                Color.PURPLE,
-                "profile is not available under the reactor model",
-                Color.END,
-            ]
-            this_msg = Color.SPACE.join(msg)
-            logger.error(this_msg)
-            self._status = -1
-            return
-        # profile data sizes
-        xsize = len(x)
-        ysize = len(y)
-        if xsize == ysize:
-            self._size = xsize
-            # independent variable (time, location, grid, ...)
-            if isinstance(x, np.double):
-                self._pos = copy.deepcopy(x)
-            else:
-                self._pos = np.array(x, dtype=np.double)
-            # dependent variable value at the corresponding position
-            if isinstance(y, np.double):
-                self._val = copy.deepcopy(y)
-            else:
-                self._val = np.array(y, dtype=np.double)
-        else:
-            msg = [
-                Color.PURPLE,
-                "the number of positions does not match the number of values\n",
-                Color.SPACEx6,
-                "number of positions =",
-                str(xsize),
-                "\n",
-                Color.SPACEx6,
-                "number of values    =",
-                str(ysize),
-                Color.END,
-            ]
-            this_msg = Color.SPACE.join(msg)
-            logger.error(this_msg)
-            self._status = -2
-
-    @property
-    def size(self) -> int:
-        """Get number of data points in the profile."""
-        """
-        Get number of data points in the profile.
-
-        Returns
-        -------
-            size: integer
-                number of profile data points
-
-        """
-        return self._size
-
-    @property
-    def status(self) -> int:
-        """Get the validity of the profile object."""
-        """
-        Get the validity of the profile object.
-
-        Returns
-        -------
-            status: integer
-                profile status
-
-        """
-        return self._status
-
-    @property
-    def pos(self) -> npt.NDArray[np.double]:
-        """Get position values of profiles data."""
-        """
-        Get position values of profiles data.
-
-        Returns
-        -------
-            pos: 1-D double array
-                position [sec, cm]
-
-        """
-        return self._pos
-
-    @property
-    def value(self) -> npt.NDArray[np.double]:
-        """Get variable values of profile data."""
-        """
-        Get variable values of profile data.
-
-        Returns
-        -------
-            val: 1-D double array
-                variable value
-
-        """
-        return self._val
-
-    @property
-    def profilekey(self) -> str:
-        """Get profile keyword."""
-        """
-        Get profile keyword.
-
-        Returns
-        -------
-            profilekeyword: string
-                keyword associated with the profile data
-
-        """
-        return self._profilekeyword
-
-    def show(self):
-        """Show the profile data."""
-        print(f"profile size: {self._size:d}")
-        print(f" position           {self._profilekeyword:s}  ")
-        for i in range(self._size):
-            print(f"{self._pos[i]:f}         {self._val[i]}")
-
-    def resetprofile(
-        self, size: int, x: npt.NDArray[np.double], y: npt.NDArray[np.double]
-    ):
-        """Reset the profile data."""
-        """
-        Reset the profile data.
-
-        Parameters
-        ----------
-            size: integer
-                number of points of the new profile data
-            x: 1-D double array
-                position of the new profile data points
-            y:1-D double array
-                variable value of the new profile data
-
-        """
-        # check array size
-        if size == self._size:
-            # new profile has the same size
-            self._pos[:] = x[:]
-            self._val[:] = y[:]
-        else:
-            # new profile has different size
-            self._size = size
-            # resize the arrays
-            self._pos.resize(size, refcheck=False)
-            self._val.resize(size, refcheck=False)
-        # fill the arrays with new values
-        self._pos[:] = x[:]
-        self._val[:] = y[:]
-
-    def getprofile_as_string_list(self) -> tuple[int, list[str]]:
-        """Create the keyword input lines as a list for Chemkin applications."""
-        """Create the keyword input lines as a list for Chemkin applications.
-
-        Returns
-        -------
-            size: integer
-                number of profile lines
-            line: list of strings
-                list of profile related keywords
-
-        """
-        # initialization
-        line = []
-        # special treatment for pressure profile
-        factor = 1.0e0
-        if self._profilekeyword == "PPRO":
-            if not Keyword.no_fullkeyword:
-                # use Full Keywords: pressure units = atm
-                factor = P_ATM
-        # assembly the profile keyword lines
-        for i in range(self._size):
-            thisline = ""
-            thisline = (
-                self._profilekeyword
-                + Keyword.fourspaces
-                + str(self._pos[i])
-                + Keyword.fourspaces
-                + str(self._val[i] / factor)
-            )
-            line.append(thisline)
-        return self._size, line
 
 
 #
@@ -775,7 +129,7 @@ class ReactorModel:
         # write text output file
         self._textout = True
         # FORTRAN file unit of the text output file
-        self._mylout = c_int(154)
+        self._mylout = ctypes.c_int(154)
         # write XML solution file
         self._xmlout = True
         # number of keywords used
@@ -832,7 +186,9 @@ class ReactorModel:
             ]
             this_msg = Color.SPACE.join(msg)
             logger.info(this_msg)
-            ierr = chemkin_wrapper.chemkin.KINInitialize(self._chemset_index, c_int(0))
+            ierr = chemkin_wrapper.chemkin.KINInitialize(
+                self._chemset_index, ctypes.c_int(0)
+            )
             if ierr == 0:
                 chemistryset_initialized(self._chemset_index.value)
             else:
@@ -1545,7 +901,7 @@ class ReactorModel:
             this_msg = Color.SPACE.join(msg)
             logger.error(this_msg)
             exit()
-        self._temperature = c_double(t)
+        self._temperature = ctypes.c_double(t)
         self.reactormixture.temperature = t
 
     @property
@@ -1579,7 +935,7 @@ class ReactorModel:
             this_msg = Color.SPACE.join(msg)
             logger.error(this_msg)
             exit()
-        self._pressure = c_double(p)
+        self._pressure = ctypes.c_double(p)
         self.reactormixture.pressure = p
 
     @property
@@ -1922,7 +1278,7 @@ class ReactorModel:
         if not mode:
             # switch to the ideal gas law
             ierr = chemkin_wrapper.chemkin.KINRealGas_UseIdealGasLaw(
-                self._chemset_index, c_int(0)
+                self._chemset_index, ctypes.c_int(0)
             )
             if ierr != 0:
                 msg = [
